requirements.txt
backoff
ctransformers
llama-cpp-python
openai
python-dotenv
